move util functions to tests/utils
