+++ tests/rpc/rpc_08-contracts_test.go
@@ -24,11 +24,12 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ChainSafe/gossamer/tests/utils"
 	"github.com/stretchr/testify/require"
 )
 
 func TestContractsRPC(t *testing.T) {
-	if GOSSAMER_INTEGRATION_TEST_MODE != rpcSuite {
+	if utils.GOSSAMER_INTEGRATION_TEST_MODE != rpcSuite {
 		_, _ = fmt.Fprintln(os.Stdout, "Going to skip RPC suite tests")
 		return
 	}
@@ -52,7 +53,7 @@ func TestContractsRPC(t *testing.T) {
 
 	t.Log("going to start gossamer")
 
-	localPidList, err := StartNodes(t, make([]*exec.Cmd, 1))
+	localPidList, err := utils.StartNodes(t, make([]*exec.Cmd, 1))
 	require.Nil(t, err)
 
 	time.Sleep(time.Second) // give server a second to start
@@ -64,11 +65,11 @@ func TestContractsRPC(t *testing.T) {
 				return
 			}
 
-			respBody, err := PostRPC(t, test.method, "http://"+GOSSAMER_NODE_HOST+":"+currentPort, "{}")
+			respBody, err := utils.PostRPC(t, test.method, "http://"+utils.GOSSAMER_NODE_HOST+":"+currentPort, "{}")
 			require.Nil(t, err)
 
 			target := reflect.New(reflect.TypeOf(test.expected)).Interface()
-			DecodeRPC(t, respBody, target)
+			utils.DecodeRPC(t, respBody, target)
 
 			require.NotNil(t, target)
 
@@ -77,6 +78,6 @@ func TestContractsRPC(t *testing.T) {
 
 	t.Log("going to TearDown Gossamer node")
 
-	errList := TearDown(t, localPidList)
+	errList := utils.TearDown(t, localPidList)
 	require.Len(t, errList, 0)
 }
